User funciotions
User functions: login, register, logout, delete, show, edit .
Entity relationships should be done but  better check.
Added to DB logo for user thinking maybe add description.
Added restrictions for example: only owner can edit and delete users and tournaments.
Can't figure out why imaged doesn't work.

+++ iis_studentske_turnaje/app/Http/Controllers/TournamentController.php
@@ -13,7 +13,7 @@ class TournamentController extends Controller
     {
         return view('tournaments.index', 
         [
-            'tournaments' => Tournament::latest()->filter(request(['search']))->paginate(1),
+            'tournaments' => Tournament::latest()->filter(request(['search']))->paginate(),
         ]);
     }
 
@@ -48,6 +48,8 @@ class TournamentController extends Controller
         if ($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
+       
+        $formFields['user_id'] = auth()->id();
 
         Tournament::create($formFields);
 
@@ -56,12 +58,22 @@ class TournamentController extends Controller
 
     // Show Edit tournament form
     public function edit(Tournament $tournament){
+        // Make sure logged in user is owner
+        if ($tournament->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         return view('tournaments.edit', ['tournament' => $tournament]);
     }
 
     // Store tournament data
     public function update(Request $request, Tournament $tournament)
     {
+        // Make sure logged in user is owner
+        if ($tournament->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'name' => ['required'],
             'game' => 'required',
@@ -83,6 +95,11 @@ class TournamentController extends Controller
 
     // Delete tournament
     public function destroy(Tournament $tournament){
+        // Make sure logged in user is owner
+        if ($tournament->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $tournament->delete();
 
         return redirect('/')
